Two 8 KB, thirty-four-arm state machines over the same object, reconstructed
in separate worktrees by sessions that could not see each other, each passing
`make phase` alone and never once compiled together.  Both function bodies
land BYTE-IDENTICAL to their branches -- 901 lines and 726 lines, diffed
against the source branches rather than asserted -- so this commit is the
prefix, the shared macro set and the record, and nothing else.

WHAT THE TWO SIDES ACTUALLY DISAGREED ABOUT

`P3D_ABS`.  Both defined it at file scope and the two comments made directly
contradictory claims about GCC 3.4.2: one said the conditional compiles to
`cltd; xor; sub`, the other said it compiles to a branch and only the builtin
gives the triple.  Settled by building this composed file both ways under the
period toolchain and comparing both symbols against the blob.  `__builtin_abs`
is closer for BOTH -- 67.2% against 65.2% on getV90Decision, 67.5% against
61.4% on getV92Decision -- and the mnemonic counts show why: where the blob has
`cltd` 6 / `neg` 4, the builtin gives 9 / 4 and the ternary gives 0 / 48.  The
effect is context-sensitive, degrading one function through `neg` and the other
through `cltd`, so a standalone `int f(int x)` would have shown the two
spellings agreeing and convicted the wrong side.  Finding 2117.

`P3D_CODE` and `P3D_SIGN` also differed and both unify on the measured
spelling; all three pairs were first proved equal in VALUE at every one of
their 130 call sites, so the choice was never behavioural.  Finding 2116.

+0x3cc is a `SerialDifferentialDecoder<int>`, +0x3f4 and +0x420 are 32-bit
fields with the same sources -- each concluded twice, independently, from
opposite ends of the class.  Finding 2118.  The member has its real type now,
so `getV92Decision`'s five casts and `t_v92dec`'s seed name it directly.

TWO THINGS FOUND WHILE COMPOSING, NEITHER CAUSED BY IT

`reanchor.py` re-anchored nine ambiguous mutations into the WRONG FUNCTION,
all nine, and reported "0 left for a human".  With no `--prefix` its score is
0 for every occurrence, so `sorted(reverse=True)` falls through to file offset
and the last match always wins -- and the guard that exists to catch exactly
that tie is gated on `args.prefix`, so it cannot fire in the one case it is
for.  Finding 347's failure mode one level up: not a lost mutation but a
RELOCATED one, which goes green while its claim has quietly changed meaning.
Repaired with `--prefix P3D_PHASE --prefix symbol` (7 placed, 2 honestly
STUCK and extended by hand).  Finding 2120.

The header's reason for accepting weak evidence at +0x3cc -- "this translation
unit is one of the fifteen the period toolchain cannot compile at all" -- is
not true, and was re-measured rather than re-read.  GCC 3.4.2 compiles this
file, at the merge base as well as here.  Finding 2119.

NOTHING WEAKENED, AND MEASURED RATHER THAN ASSERTED

  period differential: 183 passed, 0 failed     (182 would be a lost binary)
  one FAIL, Psd::process, the declared tools/gccdiverge.json entry
  refcheck 5265 refs, 0 dangling; anchorcheck 5957 mutations, 0 ambiguous
  v90p3ddec  20: 18 caught, 1 NOT caught, 1 equivalent
  v90p3dctor 23: 23 caught
  v90p3dreset 29: 28 caught, 1 NOT caught

all three identical to the summaries the v90 branch recorded, so the two
pre-existing uncaught mutations are pre-existing and the re-anchoring cost
nothing.  coverage.md regenerated for the union rather than hand-merged:
49.3% -> 51.6%, 379256 bytes, 905 symbols.

+++ include/dsplib/V90Parameters.h
@@ -410,7 +410,15 @@ public:
 	int	unnamed_434;		/* +0x434  setToDefault only; the object stores 250.0f (0x437a0000) -- 878 */
 	float	PHASE4_MEAN_ERROR_BEF_TO_AFT_UPDATE_RATIO_THRESH;	/* +0x438 */
 	float	QC_PHASE4_MEAN_ERROR_BEF_TO_AFT_UPDATE_RATIO_THRESH;	/* +0x43c */
-	int	unnamed_440;		/* +0x440  setToDefault only; the object stores 10.0f (0x41200000) -- 878 */
+	/*
+	 * A FLOAT, and it was typed `int` here only because 878 could see
+	 * the bit pattern and not the use.  `V90Phase3Demodulator::getV90Decision`
+	 * copies this word into +0x438 above -- a `float` -- with a plain
+	 * `mov`, which is what GCC emits for a float-to-float assignment and
+	 * never for an int-to-float one.  `setToDefault` stores the same
+	 * four bytes either way.
+	 */
+	float	unnamed_440;		/* +0x440  setToDefault only; 10.0f -- 878 */
 	int  	ENABLE_RRN_UP;	/* +0x444 */
 	int  	ENABLE_RRN_DOWN;	/* +0x448 */
 	int  	RATE_UP_DETECT_DURATION;	/* +0x44c */
